+++ release-notes.md
@@ -5,6 +5,8 @@ Changes:
 - Added a new `Indexer.EnableVerboseLogs` configuration to display the file/folder 
   currently being processed by the indexer.
 - Added more logs to show the read/write bandwidth
+- Take into account the "DeidentifyLogs" configuration in Orthanc to remove PHI
+  from logs.
 
 Fixes:
 - Fix #3: Added support for `InstitutionName` in the `NamingScheme`